backend/requirements.txt
flask
PyGithub
python-dotenv
flask-cors
google-generativeai